Refactor KNNVectorFieldType from KNNVectorFieldMapper to a separate class for better readability.
