+++ src/main/java/org/opensearch/knn/index/mapper/KNNVectorFieldMapperUtil.java
@@ -236,7 +236,7 @@ public class KNNVectorFieldMapperUtil {
      * @param knnVectorFieldType knn vector field type
      * @return expected vector length
      */
-    public static int getExpectedVectorLength(final KNNVectorFieldMapper.KNNVectorFieldType knnVectorFieldType) {
+    public static int getExpectedVectorLength(final KNNVectorFieldType knnVectorFieldType) {
         int expectedDimensions = knnVectorFieldType.getDimension();
         if (isModelBasedIndex(expectedDimensions)) {
             ModelMetadata modelMetadata = getModelMetadataForField(knnVectorFieldType);
@@ -255,7 +255,7 @@ public class KNNVectorFieldMapperUtil {
      * @param knnVectorField knn vector field
      * @return the model metadata from knnVectorField
      */
-    private static ModelMetadata getModelMetadataForField(final KNNVectorFieldMapper.KNNVectorFieldType knnVectorField) {
+    private static ModelMetadata getModelMetadataForField(final KNNVectorFieldType knnVectorField) {
         String modelId = knnVectorField.getModelId();
 
         if (modelId == null) {
